+++ mathematica/SumsOfPowersViaCentralFiniteDifferencesAndNewtonFormula.m
@@ -58,6 +58,10 @@ DoubleCenteredSumsOfPowers::usage=""
 
 ValidateDoubleCenteredSumsOfPowers::usage=""
 
+MultifoldCenteredSumsOfPowers::usage=""
+
+ValidateMultifoldCenteredSumsOfPowers::usage=""
+
 (*END: Definitions *)
 
 (* =========================================================================DOCS END=================================================================== *)
@@ -205,6 +209,26 @@ DoubleCenteredSumsOfPowers[n_, m_, t_] :=
   ];
 
 ValidateDoubleCenteredSumsOfPowers[max_] := Table[MultifoldSumOfPowersRecurrence[2, n, m] - DoubleCenteredSumsOfPowers[n, m, t], {n, 0, max}, {m, 0, max}, {t, 0, max}] //Flatten
+
+MultifoldCenteredSumsOfPowers[r_, n_, m_, t_] :=
+  t^m*MultifoldSumOfPowersRecurrence[r, n, 0] +
+  Sum[
+    CentralDifference[t, m, k]/2 *
+      (
+        (Binomial[n - t + k/2 + r, k + r] +
+         Binomial[n - t + k/2 + r - 1, k + r])
+        -
+        Sum[
+          (Binomial[-t + k/2 + r - s, k + r - s] +
+           Binomial[-t + k/2 + r - s - 1, k + r - s])*
+            MultifoldSumOfPowersRecurrence[s, n, 0],
+          {s, 0, r - 1}
+        ]
+      ),
+    {k, 1, m}
+  ];
+
+ValidateMultifoldCenteredSumsOfPowers[max_] := Table[MultifoldSumOfPowersRecurrence[r, n, m] - MultifoldCenteredSumsOfPowers[r, n, m, t], {n, 0, max}, {m, 0, max}, {t, 0, max}, {r, 0, max}] //Flatten
 
 (*END: Definitions *)