+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Add simplified formula for centered sums
 - Add formula for double centered sums
 - Add Mathematica programs
+- Add formula for multifold sums of powers
 
 ## [1.4.0] - 18-Jan-2026